% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_addressbase_plus_single_line_address.R
\name{calc_addressbase_plus_geo_single_line_address}
\alias{calc_addressbase_plus_geo_single_line_address}
\title{Calculate AddressBase Plus GEO single line address}
\usage{
calc_addressbase_plus_geo_single_line_address(df, include_postcode = FALSE)
}
\arguments{
\item{df}{AddressBase Plus DB table}

\item{include_postcode}{Whether or not to include postcode. Default is FALSE.}
}
\description{
Calculate AddressBase Plus GEO single line address.
}
