% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_match_addresses.R
\name{calc_match_addresses}
\alias{calc_match_addresses}
\title{Match two sets of addresses}
\usage{
calc_match_addresses(
  primary_df,
  primary_postcode_col,
  primary_address_col,
  lookup_df,
  lookup_postcode_col,
  lookup_address_col
)
}
\arguments{
\item{primary_df}{Dataframe of distinct primary addresses}

\item{primary_postcode_col}{Column containing the primary postcode}

\item{primary_address_col}{Column containing the primary single line address}

\item{lookup_df}{Dataframe of distinct lookup addresses}

\item{lookup_postcode_col}{Column containing the lookup postcode}

\item{lookup_address_col}{Column containing the lookup single line address}
}
\description{
Match a distinct dataframe of primary addresses to a distinct dataframe of
lookup addresses.
}
\details{
Returns a dataframe of exact matches (postcode and single line address) and
non exact matches (postcode and fuzzy single line address). For non exact
matches it will retain draws.
}
