% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_os_data_package.R
\name{download_os_data_package}
\alias{download_os_data_package}
\title{Download a data package from Ordinance Survey Data Hub}
\usage{
download_os_data_package(
  key = Sys.getenv("OS_API_KEY"),
  data_package_id,
  data_package_version = NA,
  destination,
  unzip = TRUE
)
}
\arguments{
\item{key}{OS Data Hub API key (with permissions for required data package.
Default is the environment variable \code{Sys.getenv("OS_API_KEY")}.}

\item{data_package_id}{Data package ID}

\item{data_package_version}{Data package version. Default if \code{NA} is most
recent.}

\item{destination}{Destination (folder if \code{unzip == TRUE} or file if
\code{unzip == FALSE}).}

\item{unzip}{Whether or not to unzip the file. Default is \code{TRUE}.}
}
\description{
Download a data package from Ordinance Survey Data Hub
}
