% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_strings_df.R
\name{merge_strings_df}
\alias{merge_strings_df}
\title{Merge Two Strings Together Whilst Retaining an Order of Some Kind}
\usage{
merge_strings_df(string1, string2)
}
\arguments{
\item{df}{local df}

\item{first_col}{First column}

\item{second_col}{Second column}

\item{merge_col}{Name to give merged column}
}
\value{
when used with mapply, original df with additional merged column
}
\description{
This is intended to be used with differing address strings
The result is a 'master' address string that contains all parts of each address
}
\examples{
table_df \%>\%
  mutate(
    merged_string = mapply(merge_strings_df, string1, string2)
    )

}
