% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oracle_create_addressbase_plus_table.R
\name{oracle_create_addressbase_plus_table}
\alias{oracle_create_addressbase_plus_table}
\title{Upload AddressBase Plus to Oracle}
\usage{
oracle_create_addressbase_plus_table(con, path, pattern = NULL)
}
\arguments{
\item{con}{DB Connection.}

\item{path}{Path to .csv files}

\item{pattern}{Pattern of files to upload. Default is NULL (all files)}
}
\description{
Upload AddressBase Plus to an Oracle DB. First download the data package for
AddressBase Plus using \code{addressMatchR::download_os_data_package()} and then
load it into the DB using this function.
}
\details{
Using guidance from
https://www.ordnancesurvey.co.uk/documents/addressbase-products-getting-started-guide2.pdf
}
