% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/oracle_create_addressbase_plus_single_line_address_table.R
\name{oracle_create_addressbase_single_line_address_table}
\alias{oracle_create_addressbase_single_line_address_table}
\title{Create a version of AddressBase Plus with the single line addresses}
\usage{
oracle_create_addressbase_single_line_address_table(
  df,
  table_name = "ADDRESS_BASE_PLUS_SINGLE_LINE_ADDRESS"
)
}
\arguments{
\item{df}{Database table ("ADDRESSBASE_PLUS")}

\item{table_name}{Table name. Default to
"ADDRESS_BASE_PLUS_SINGLE_LINE_ADDRESS"}
}
\description{
Create a processed version of AddressBase Plus that contains the main columns
needed along with up to 3 versions of a single line address per UPRN.
}
