% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_postcode_df.R
\name{tidy_postcode_df}
\alias{tidy_postcode_df}
\title{Tidy a datarame postcode}
\usage{
tidy_postcode_df(df, col)
}
\arguments{
\item{df}{dataFrame}

\item{col}{Postcode column}
}
\description{
Tidy a postcode so it is ready for joining.
}
