% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_single_line_address_df.R
\name{tidy_single_line_address_df}
\alias{tidy_single_line_address_df}
\title{Tidy a dataframe single line address}
\usage{
tidy_single_line_address_df(df, col, remove_postcode = FALSE)
}
\arguments{
\item{df}{Must be a dataframe (i.e. not a lazy_frame)}

\item{col}{Single line address column}

\item{remove_postcode}{If to remove the postcode. Default is FALSE}
}
\description{
Tidy a single line address ready for tokenising.
}
