% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_addressbase_plus_single_line_address.R
\name{calc_addressbase_plus_geo_single_line_address}
\alias{calc_addressbase_plus_geo_single_line_address}
\title{Calculate AddressBase Plus GEO single line address}
\usage{
calc_addressbase_plus_geo_single_line_address(df)
}
\arguments{
\item{df}{AddressBase Plus DB table}
}
\description{
Calculate AddressBase Plus GEO single line address.
}
\examples{

}
