% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_postcode.R
\name{tidy_postcode}
\alias{tidy_postcode}
\title{Tidy a postcode}
\usage{
tidy_postcode(df, col)
}
\arguments{
\item{df}{Database table}

\item{col}{Postcode column}
}
\description{
Tidy a postcode so it is ready for joining.
}
